codecov: don't generate code coverage if the tests failed

+++ internal/testing/runchecks.sh
@@ -32,7 +32,7 @@ result=0
 # because it is slow, and codecov will only save the last one anyway.
 if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then
   go test -race -coverpkg=./... -coverprofile=coverage.out ./... || result=1
-  if [ -f coverage.out ]; then
+  if [ -f coverage.out ] && [ $result -eq 0 ]; then
     # Filter out test and sample packages.
     grep -v test coverage.out | grep -v samples > coverage2.out
     mv coverage2.out coverage.out
